+++ old examples.py
